abstract shader descriptor structs, FileSystem almost finished
Work in progress of a descriptor struct-based shader initialization idea, to be easy to make shaders in a similar manner D3D11 classes are initialized; it's kinda finished, but I need to give it a final test to it. Finished FileSystem class, although I need to add method descriptions.
Also I changed ConsoleLogger::consolePrint() method to simple ConsoleLogger::Print()